create voting verifier api
